Isolated test halo generation into separate file. Added documentation.
